UPDATE chartsCO2 (while API not responding)
